Fix: Remove invalid collections field from meta/main.yml
The 'collections:' field is not valid in meta/main.yml and was causing:
'ERROR! Unexpected Exception, this is probably a bug: unhashable type: dict'

Collection dependencies should only be specified in requirements.yml file,
not in the role metadata. The meta/main.yml file only supports 'dependencies:'
for role dependencies, not collection dependencies.